fixed create table

+++ api/model/customer_model.py
@@ -9,7 +9,7 @@ class customerModel(Based, baseModel):
     __tablename__ = "customer"
        
     customer_name = Column(String(64))
-    customer_phone = Column(Integer(13))
+    customer_phone = Column(Integer)
     
     #one-to-many
     orders = relationship("orderModel", backref="customerModel", lazy=False, cascade="all, delete")
