+++ api/model/supplier_model.py
@@ -12,7 +12,7 @@ class supplierModel(Based, baseModel):
     supplier_name = Column(String(128), nullable= False)
     supplier_desc = Column(String(512))
     supplier_contact = Column(String(64))
-    supplier_phone = Column(Integer(13))
+    supplier_phone = Column(Integer)
     supplier_address = Column(String(512))
     supplier_city = Column(String(64))
     supplier_district = Column(String(64))
